OrdersVC: Move presentDetails to MasterVC
